src/hybrid_color_correction/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=9.5.0
streamlit>=1.28.0
einops>=0.7.0
numpy>=1.24.0
